pyproject.toml
[project]
name = "myfit-api"
version = "1.0.0"
description = "MyFit Platform API - White-label fitness platform backend"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Marcel Piva", email = "marcel@myfitplatform.com" }
]

dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "gunicorn>=23.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Redis
    "redis>=5.2.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",

    # Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",

    # Background Tasks
    "celery>=5.4.0",

    # HTTP Client
    "httpx>=0.28.0",

    # AI Services
    "openai>=1.57.0",
    "anthropic>=0.40.0",

    # File Storage
    "boto3>=1.35.0",

    # i18n
    "babel>=2.16.0",

    # WebSocket
    "websockets>=14.1",

    # Utils
    "python-dotenv>=1.0.1",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",

    # API Documentation
    "scalar-fastapi>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501", "B008", "C901"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
